+++ cmd/hidden_lake/applications/messenger/pkg/app/app_test.go
@@ -42,7 +42,6 @@ func TestApp(t *testing.T) {
 		FSettings: &config.SConfigSettings{
 			FMessagesCapacity: testutils.TCCapacity,
 			FWorkSizeBits:     testutils.TCWorkSize,
-			FPseudonym:        "test_pseudonym",
 		},
 		FAddress: &config.SAddress{
 			FInterface: testutils.TgAddrs[57],